+++ Engine/Core/ImGui/ImGuiSample.h
@@ -1,0 +1,17 @@
+// ImGuiSample.h
+// created on 2021/6/12
+// author @zoloypzuo
+
+#pragma once
+
+#include "ZeloPrerequisites.h"
+
+namespace ImGui {
+
+void ShowUserGuide();
+
+void ShowStyleEditor(ImGuiStyle *ref = NULL);
+
+void ShowTestWindow(bool *open = NULL);
+
+}
